fix functional test runner
